Move NuGet configuration test to NuGet integration tests. (#715)
