+++ nuget_test.go
@@ -1,10 +1,14 @@
 package main
 
 import (
-	"github.com/jfrog/jfrog-cli/artifactory/utils/dotnet"
+	"encoding/xml"
+	"github.com/jfrog/jfrog-cli/artifactory/commands/dotnet"
+	dotnetutils "github.com/jfrog/jfrog-cli/artifactory/utils/dotnet"
 	"github.com/jfrog/jfrog-cli/utils/cliutils"
+	"github.com/jfrog/jfrog-cli/utils/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -52,21 +56,21 @@ func TestNugetResolve(t *testing.T) {
 
 func TestNativeNugetResolve(t *testing.T) {
 	uniqueNugetTests := []testDescriptor{
-		{"packagesconfigwithoutmodulechnage", "packagesconfig", []string{dotnet.Nuget.String(), "restore"}, []string{"packagesconfig"}, []int{6}},
-		{"packagesconfigwithmodulechnage", "packagesconfig", []string{dotnet.Nuget.String(), "restore", "--module=" + ModuleNameJFrogTest}, []string{ModuleNameJFrogTest}, []int{6}},
-		{"packagesconfigwithconfigpath", "packagesconfig", []string{dotnet.Nuget.String(), "restore", "./packages.config", "-SolutionDirectory", "."}, []string{"packagesconfig"}, []int{6}},
-		{"multipackagesconfigwithoutmodulechnage", "multipackagesconfig", []string{dotnet.Nuget.String(), "restore"}, []string{"proj1", "proj2"}, []int{4, 3}},
-		{"multipackagesconfigwithmodulechnage", "multipackagesconfig", []string{dotnet.Nuget.String(), "restore", "--module=" + ModuleNameJFrogTest}, []string{ModuleNameJFrogTest}, []int{6}},
-		{"multipackagesconfigwithslnPath", "multipackagesconfig", []string{dotnet.Nuget.String(), "restore", "./multipackagesconfig.sln"}, []string{"proj1", "proj2"}, []int{4, 3}},
-		{"multipackagesconfigsingleprojectdir", "multipackagesconfig", []string{dotnet.Nuget.String(), "restore", "./proj2/", "-SolutionDirectory", "."}, []string{"proj2"}, []int{3}},
-		{"multipackagesconfigsingleprojectconfig", "multipackagesconfig", []string{dotnet.Nuget.String(), "restore", "./proj1/packages.config", "-SolutionDirectory", "."}, []string{"proj1"}, []int{4}},
+		{"packagesconfigwithoutmodulechnage", "packagesconfig", []string{dotnetutils.Nuget.String(), "restore"}, []string{"packagesconfig"}, []int{6}},
+		{"packagesconfigwithmodulechnage", "packagesconfig", []string{dotnetutils.Nuget.String(), "restore", "--module=" + ModuleNameJFrogTest}, []string{ModuleNameJFrogTest}, []int{6}},
+		{"packagesconfigwithconfigpath", "packagesconfig", []string{dotnetutils.Nuget.String(), "restore", "./packages.config", "-SolutionDirectory", "."}, []string{"packagesconfig"}, []int{6}},
+		{"multipackagesconfigwithoutmodulechnage", "multipackagesconfig", []string{dotnetutils.Nuget.String(), "restore"}, []string{"proj1", "proj2"}, []int{4, 3}},
+		{"multipackagesconfigwithmodulechnage", "multipackagesconfig", []string{dotnetutils.Nuget.String(), "restore", "--module=" + ModuleNameJFrogTest}, []string{ModuleNameJFrogTest}, []int{6}},
+		{"multipackagesconfigwithslnPath", "multipackagesconfig", []string{dotnetutils.Nuget.String(), "restore", "./multipackagesconfig.sln"}, []string{"proj1", "proj2"}, []int{4, 3}},
+		{"multipackagesconfigsingleprojectdir", "multipackagesconfig", []string{dotnetutils.Nuget.String(), "restore", "./proj2/", "-SolutionDirectory", "."}, []string{"proj2"}, []int{3}},
+		{"multipackagesconfigsingleprojectconfig", "multipackagesconfig", []string{dotnetutils.Nuget.String(), "restore", "./proj1/packages.config", "-SolutionDirectory", "."}, []string{"proj1"}, []int{4}},
 	}
 	testNativeNugetDotnetResolve(t, uniqueNugetTests, tests.NugetBuildName, utils.Nuget)
 }
 
 func TestDotnetResolve(t *testing.T) {
 	uniqueDotnetTests := []testDescriptor{
-		{"multireferencesingleprojectdir", "multireference", []string{dotnet.DotnetCore.String(), "restore", "src/multireference.proj1/"}, []string{"proj1"}, []int{5}},
+		{"multireferencesingleprojectdir", "multireference", []string{dotnetutils.DotnetCore.String(), "restore", "src/multireference.proj1/"}, []string{"proj1"}, []int{5}},
 	}
 	testNativeNugetDotnetResolve(t, uniqueDotnetTests, tests.DotnetBuildName, utils.Dotnet)
 }
@@ -145,4 +149,79 @@ func runNuGet(t *testing.T, args ...string) {
 	artifactoryNuGetCli := tests.NewJfrogCli(execMain, "jfrog rt", "")
 	err := artifactoryNuGetCli.Exec(args...)
 	assert.NoError(t, err)
+}
+
+func TestInitNewConfig(t *testing.T) {
+	initNugetTest(t)
+
+	tempDirPath, err := fileutils.CreateTempDir()
+	if err != nil {
+		t.Error(err)
+	}
+	defer fileutils.RemoveTempDir(tempDirPath)
+
+	c := &dotnetutils.Cmd{}
+	params := &dotnet.DotnetCommand{}
+	params.SetRtDetails(&config.ArtifactoryDetails{Url: "http://some/url", User: "user", Password: "password"})
+	configFile, err := dotnet.WriteToTempConfigFile(c, tempDirPath)
+	if err != nil {
+		t.Error(err)
+	}
+
+	// Prepare the config file with NuGet authentication
+	err = params.AddNugetAuthToConfig(dotnetutils.Nuget, configFile)
+	if err != nil {
+		t.Error(err)
+	}
+
+	content, err := ioutil.ReadFile(configFile.Name())
+	if err != nil {
+		t.Error(err)
+	}
+
+	nugetConfig := NugetConfig{}
+	err = xml.Unmarshal(content, &nugetConfig)
+	if err != nil {
+		t.Error("Unmarshalling failed with an error:", err.Error())
+	}
+
+	if len(nugetConfig.PackageSources) != 1 {
+		t.Error("Expected one package sources, got", len(nugetConfig.PackageSources))
+	}
+
+	source := "http://some/url/api/nuget"
+
+	for _, packageSource := range nugetConfig.PackageSources {
+		if packageSource.Key != dotnet.SourceName {
+			t.Error("Expected", dotnet.SourceName, ",got", packageSource.Key)
+		}
+
+		if packageSource.Value != source {
+			t.Error("Expected", source, ", got", packageSource.Value)
+		}
+	}
+
+	if len(nugetConfig.PackageSourceCredentials) != 1 {
+		t.Error("Expected one packageSourceCredentials, got", len(nugetConfig.PackageSourceCredentials))
+	}
+
+	if len(nugetConfig.PackageSourceCredentials[0].JFrogCli) != 2 {
+		t.Error("Expected two fields in the JFrogCli credentials, got", len(nugetConfig.PackageSourceCredentials[0].JFrogCli))
+	}
+}
+
+type NugetConfig struct {
+	XMLName                  xml.Name                   `xml:"configuration"`
+	PackageSources           []PackageSources           `xml:"packageSources>add"`
+	PackageSourceCredentials []PackageSourceCredentials `xml:"packageSourceCredentials"`
+	Apikeys                  []PackageSources           `xml:"apikeys>add"`
+}
+
+type PackageSources struct {
+	Key   string `xml:"key,attr"`
+	Value string `xml:"value,attr"`
+}
+
+type PackageSourceCredentials struct {
+	JFrogCli []PackageSources `xml:">add"`
 }
